+++ r-package/tests/testthat/test_read_grid.R
@@ -5,8 +5,8 @@ context("read_grid")
 test_that("read_grid expected behavior", {
 
   # whole file
-  expect_true(is(  read_grid(city='nat'), 'sf'))
-  expect_true(is(  read_grid(city='Natal'), 'sf'))
+  expect_true(is(  read_grid(city='nat', showProgress = FALSE), 'sf'))
+  expect_true(is(  read_grid(city='Natal', showProgress = FALSE), 'sf'))
 
 })