+++ r-package/tests/testthat/test_read_landuse.R
@@ -5,8 +5,8 @@ context("read_landuse")
   test_that("read_landuse expected behavior", {
 
     # whole file
-    expect_true(is(  read_landuse(city='nat', geometry = TRUE), 'sf'))
-    expect_true(is(  read_landuse(city='nat', geometry = FALSE), 'data.frame'))
+    expect_true(is(  read_landuse(city='nat', geometry = TRUE, showProgress = FALSE), 'sf'))
+    expect_true(is(  read_landuse(city='nat', geometry = FALSE, showProgress = FALSE), 'data.frame'))
     expect_true(is(  read_landuse(city='nat', geometry = FALSE, showProgress = FALSE), 'data.frame'))
   })